+++ packages/sugar-server/src/core/application.ts
@@ -1,4 +1,4 @@
-import { IncomingMessage, ServerResponse } from 'http';
+import { IncomingMessage, Server, ServerResponse } from 'http';
 import url from 'url';
 import Koa from 'koa';