+++ packages/sugar-server/src/shared/logger.ts
@@ -1,0 +1,19 @@
+export const success = (message: string) => {
+  console.log(`✓ ${message}`);
+}
+
+export const error = (message: string) => {
+  console.log(`✘ ${message}`);
+}
+
+export const warn = (message: string) => {
+  console.log(`⚠︎ ${message}`);
+}
+
+export const info = (message: string) => {
+  console.log(`➤ ${message}`);
+}
+
+export const log = (message: string) => {
+  console.log(message);
+}